--- SuspendedLocation/src/main/java/com/xmartlabs/suspendedlocation/LatLong.kt
@@ -1,3 +1,0 @@
-package com.xmartlabs.suspendedlocation
-
-data class LatLong(val lat: Double, val long: Double)
