--- SuspendedPlaces/build.gradle
@@ -1,41 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-
-android {
-    compileSdkVersion androidVersions.compileSdk
-    buildToolsVersion androidVersions.buildTools
-
-    defaultConfig {
-        minSdkVersion androidVersions.minSdk
-        targetSdkVersion androidVersions.targetSdk
-        versionCode 1
-        versionName ""
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: "libs", include: ["*.jar"])
-    implementation project(":SuspendedLocation")
-
-    implementation deps.anotations
-    implementation deps.appcompat
-    implementation deps.coreKtx
-    implementation deps.corutines
-    implementation deps.kotlin
-    implementation deps.places
-    implementation deps.playServicesLocation
-
-    testImplementation test.junit
-    androidTestImplementation test.testJunit
-    androidTestImplementation test.espresso
-}
