--- SuspendedPlaces/src/main/java/com/xmartlabs/suspendedplaces/PlacesNotInitializedException.kt
@@ -1,10 +1,0 @@
-package com.xmartlabs.suspendedplaces
-
-import java.lang.RuntimeException
-
-/*
-* Places SDK must be initialized before using its functions for more info follow up this link
-* https://developers.google.com/places/android-sdk/start#maps_places_get_started-kotlin
-*
-*  */
-class PlacesNotInitializedException : RuntimeException("Places not initialized")
