--- SuspendedPlaces/src/main/java/com/xmartlabs/suspendedplaces/TaskExtensions.kt
@@ -1,15 +1,0 @@
-package com.xmartlabs.suspendedplaces
-
-import com.google.android.gms.tasks.Task
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-import kotlin.coroutines.suspendCoroutine
-
-suspend fun <T> Task<T>.taskToCoroutine() = suspendCoroutine<T> { continuation ->
-    addOnSuccessListener { result ->
-        continuation.resume(result)
-    }
-    addOnFailureListener { exception ->
-        continuation.resumeWithException(exception)
-    }
-}
