+++ core/src/androidTest/java/com/xmartlabs/suspendedlocation/ExampleInstrumentedTest.kt
