+++ core/src/main/java/com/xmartlabs/suspendedlocation/core/LatLong.kt
@@ -1,0 +1,3 @@
+package com.xmartlabs.suspendedlocation.core
+
+data class LatLong(val lat: Double, val long: Double)
