+++ places/.gitignore
