+++ places/src/main/java/com/xmartlabs/suspendedlocation/places/TaskExtensions.kt
@@ -1,0 +1,15 @@
+package com.xmartlabs.suspendedlocation.places
+
+import com.google.android.gms.tasks.Task
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
+import kotlin.coroutines.suspendCoroutine
+
+suspend fun <T> Task<T>.taskToCoroutine() = suspendCoroutine<T> { continuation ->
+    addOnSuccessListener { result ->
+        continuation.resume(result)
+    }
+    addOnFailureListener { exception ->
+        continuation.resumeWithException(exception)
+    }
+}
